move common stuff into concerns
