+++ lib/mhv_logging/client.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 require 'faraday'
 require 'common/client/errors'
+require 'common/client/concerns/client_methods'
+require 'common/client/concerns/session_based_client'
 require 'common/client/middleware/response/json_parser'
 require 'common/client/middleware/response/raise_error'
 require 'common/client/middleware/response/snakecase'
@@ -16,60 +18,14 @@ require 'rx/client_session'
 module MHVLogging
   # Core class responsible for api interface operations
   class Client
+    include Common::ClientMethods
+    include Common::SessionBasedClient
     include MHVLogging::API::Audits
     include Rx::API::Sessions
-
-    attr_reader :config, :session
 
     def initialize(session:)
       @config = Rx::Configuration.instance
       @session = Rx::ClientSession.find_or_build(session)
-    end
-
-    # Note this uses the same session store as Rx
-    def authenticate
-      if @session.expired?
-        @session = get_session
-        @session.save
-      end
-      self
-    end
-
-    private
-
-    def perform(method, path, params, headers = nil)
-      raise NoMethodError, "#{method} not implemented" unless config.request_types.include?(method)
-
-      send(method, path, params || {}, headers)
-    end
-
-    def request(method, path, params = {}, headers = {})
-      raise_not_authenticated if headers.keys.include?('Token') && headers['Token'].nil?
-      connection.send(method.to_sym, path, params) { |request| request.headers.update(headers) }.env
-    rescue Faraday::Error::TimeoutError, Timeout::Error => error
-      raise Common::Client::Errors::RequestTimeout, error
-    rescue Faraday::Error::ClientError => error
-      raise Common::Client::Errors::Client, error
-    end
-
-    def get(path, params, headers = base_headers)
-      request(:get, path, params, headers)
-    end
-
-    def post(path, params, headers = base_headers)
-      request(:post, path, params, headers)
-    end
-
-    def raise_not_authenticated
-      raise Common::Client::Errors::NotAuthenticated, 'Not Authenticated'
-    end
-
-    def auth_headers
-      config.base_request_headers.merge('appToken' => config.app_token, 'mhvCorrelationId' => @session.user_id.to_s)
-    end
-
-    def token_headers
-      config.base_request_headers.merge('Token' => @session.token)
     end
 
     def connection
